Add AppVersion and BuildID, replace Start Text Type menu with hotkey
Implemented AppVersion and BuildID for better tracking and replaced "Start Text Type" menu with a hotkey for efficiency. The 'textType' function now gets invoked by a hotkey press rather than a menu click. This change streamlines the usage of the application. Furthermore, the function for typing text has been updated to use 'robotgo' which is more efficient and reliable for automating keystrokes. Also, setup of hotkey and clipboard has been modularized to the 'setup' function.