Update hotkeys and automation for better efficiency and reliability
This revision adds the implementation of AppVersion and BuildID for better version control. The "Start Text Type" menu is replaced with a hotkey feature for improved efficiency when triggering text input. Leveraged 'robotgo' for text typing function to automate the keystrokes more efficiently. Additionally, the setup of hotkey and clipboard operations is now modularized into a 'setup' function for better code organization and improved readability. This change aims to enhance the overall user experience by increasing the application's reliability and performance.