Update README with new feature and minor text changes
This commit includes two key updates in the README file. Firstly, the functionality of password entry into console applications now supported in the app has been documented, detailing the specific hotkeys for two distinct keyboard layouts. This addition is important for user awareness and usability of the new feature. Secondly, a minor linguistic change was made in the 'Compilation' section for clarity and consistency in the instructions on how to compile the program.