Update README with correct build instructions and usage details
This commit fixes a typo in the Compilation section of the README ('INSERT_VERSION' is corrected to 'VERSION') and append details about the acquisition and execution of the application. Instructions on altering the hotkey were added as well. This modification makes for a smoother end-user experience by providing clear directions for program operation post-installation.