Add EnterKey functionality to Text Type application
The update includes the addition of a new feature in the Text Type application, which lets the user decide if they want the 'Enter' key to be automatically pressed after the text is pasted from the clipboard. This change increases the flexibility and accommodate various user text input needs. The product and file version in versioninfo.json has also been incremented from 1.1.0 to 1.1.1 to indicate the addition of this new functionality.