Add checks for application version updates
Added a new function called "checkUpdate" to the main.go file. This function aims to check automatically if the current version of the application is outdated by comparing it with the latest release on the specified GitHub repository. If the application is outdated, an alert will be shown to the user and the user will be redirected to the latest release on the GitHub repository. This will help in keeping the application up to date with the newest features and improvements. Also, there are three new packages imported to handle alerts, browser activities and version comparison in a more sophisticated and streamlined manner.