Update version number in versioninfo.json file
Bumped up the patch version number from 1.1.1 to 1.1.2 in versioninfo.json to reflect the implementation of a new feature. This change was needed to implement automatic checks for application updates, ensuring our users are always aware of the latest versions and updates. These changes will accommodate the introduction of alerts and redirection to the updated version for the users.