Add version check functionality in Go-TextType app
This commit includes a new package, 'update', in Go-TextType application where a function 'CheckUpdate' is added to check if there's a new version of Go-TextType available on GitHub. It compares current 'AppVersion' with the latest one on GitHub. If a new version found, it shows an alert with the option to download the latest release. In case of any error while performing operations, it calls 'CheckError' from 'errorHandling' package. A helper function 'deleteFile' is also added to delete temporary dump files. This enhancement provides the straightforward way for users to make sure they are using the latest version of the app.