Refactor main.go to modularize repeated code
This update refactors main.go in the Go-TextType app. Major changes occurred primarily through moving sections of the code into separate files inside an internal folder, resulting in the importation of several new files. These changes were made to enhance the organization and readability of the code by removing duplicated code and replacing it with function calls to the respective modules. This approach also helps with code maintenance and will improve productivity when updating or debugging the software.