Add build, clean, and dep functions to make.ps1
In this commit, a new make.ps1 file is created which includes 'build', 'clean', and 'dep' functions. The 'build' function makes a new version of the project, including generating a binary icon and version info, and building the project. The 'clean' function cleans the project, including deleting the binary folder. The 'dep' function installs dependencies. These changes are useful for automating build process.