Update README.md with important notices and revised compilation steps
This commit modifies the README.md to provide users with more useful and up-to-date information. The alert box has been marked as 'Important' instead of 'Note', to grab users' attention. Also, the option to contact for additional Hotkeys is extended to reporting issues. The 'Dependencies' section explains the purpose of new Go packages for update notifications, opening a browser, and accessing the latest version. The 'Compilation' instructions are expanded for clarity, including steps to clean the project area and install dependencies. These changes enhance end user experience and improve the ease of project setup for developers.