Remove starter-topic.md and update documentation configuration
This commit includes the removal of the starter-topic.md file as it was no longer needed. It was initially used as a template file for the new documentation, but its purpose has been served and is now irrelevant. The writerside.cfg file has also been modified to reflect the version number change to v1.1.3, and a new code-snippets directory is added. Also, buildprofiles.xml now has a more descriptive build profile with added attributes for color presets, copyright and social media links. A new file, redirection-rules.xml, is introduced to handle redirection rules for the deleted page. Finally, a comprehensive README.md file has been included in the topics, replacing the starter-topic.md as the index file, providing more relevant content. Overall, these changes improve the organization, relevancy and navigability of the project's documentation.