Add CONTRIBUTING.md and simplify README.md
This commit introduces a CONTRIBUTING.md guide to outline steps for issue reporting and contributing patches. Including this guide supports an open-source community and brings clarity to the contribution process. Also, the README.md was simplified by removing detailed information on dependencies and compilation instructions and replacing with a reference to a detailed guide in CONTRIBUTING.md. This makes the README.md file more compact and reader-friendly while maintaining all the necessary information for those who are interested in contributing or compiling the application.